+++ aiven/resource_elasticsearch.go
@@ -5,6 +5,8 @@ package aiven
 import (
 	"time"
 
+	"github.com/aiven/terraform-provider-aiven/pkg/service"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/customdiff"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
@@ -37,7 +39,17 @@ func resourceElasticsearch() *schema.Resource {
 		ReadContext:   resourceServiceRead,
 		UpdateContext: resourceServiceUpdate,
 		DeleteContext: resourceServiceDelete,
-		CustomizeDiff: resourceServiceCustomizeDiffWrapper(ServiceTypeElasticsearch),
+		CustomizeDiff: customdiff.All(
+			customdiff.Sequence(
+				service.SetServiceTypeIfEmpty(ServiceTypeElasticsearch),
+				customdiff.IfValueChange("disk_space",
+					service.DiskSpaceShouldNotBeEmpty,
+					service.CustomizeDiffCheckDiskSpace),
+			),
+			customdiff.IfValueChange("service_integrations",
+				service.ServiceIntegrationShouldNotBeEmpty,
+				service.CustomizeDiffServiceIntegrationAfterCreation),
+		),
 		Importer: &schema.ResourceImporter{
 			StateContext: resourceServiceState,
 		},