+++ aiven/resource_kafka.go
@@ -4,15 +4,24 @@ package aiven
 
 import (
 	"context"
+	"strconv"
 	"time"
 
 	"github.com/aiven/aiven-go-client"
+	"github.com/aiven/terraform-provider-aiven/pkg/service"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/customdiff"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
 func aivenKafkaSchema() map[string]*schema.Schema {
 	aivenKafkaSchema := serviceCommonSchema()
+	aivenKafkaSchema["karapace"] = &schema.Schema{
+		Type:             schema.TypeBool,
+		Optional:         true,
+		Description:      "Switch the service to use Karapace for schema registry and REST proxy",
+		DiffSuppressFunc: emptyObjectDiffSuppressFunc,
+	}
 	aivenKafkaSchema["default_acl"] = &schema.Schema{
 		Type:        schema.TypeBool,
 		Optional:    true,
@@ -74,10 +83,9 @@ func resourceKafka() *schema.Resource {
 	return &schema.Resource{
 		Description:   "The Kafka resource allows the creation and management of Aiven Kafka services.",
 		CreateContext: resourceKafkaCreate,
-		ReadContext:   resourceServiceRead,
+		ReadContext:   resourceKafkaRead,
 		UpdateContext: resourceServiceUpdate,
 		DeleteContext: resourceServiceDelete,
-		CustomizeDiff: resourceServiceCustomizeDiffWrapper(ServiceTypeKafka),
 		Importer: &schema.ResourceImporter{
 			StateContext: resourceServiceState,
 		},
@@ -87,6 +95,39 @@ func resourceKafka() *schema.Resource {
 		},
 
 		Schema: aivenKafkaSchema(),
+		CustomizeDiff: customdiff.All(
+			customdiff.Sequence(
+				service.SetServiceTypeIfEmpty(ServiceTypeKafka),
+				customdiff.IfValueChange("disk_space",
+					service.DiskSpaceShouldNotBeEmpty,
+					service.CustomizeDiffCheckDiskSpace),
+			),
+			customdiff.IfValueChange("service_integrations",
+				service.ServiceIntegrationShouldNotBeEmpty,
+				service.CustomizeDiffServiceIntegrationAfterCreation),
+
+			// if a kafka_version is >= 3.0 then this schema field is not applicable
+			customdiff.ComputedIf("karapace", func(ctx context.Context, d *schema.ResourceDiff, m interface{}) bool {
+				project := d.Get("project").(string)
+				serviceName := d.Get("service_name").(string)
+				client := m.(*aiven.Client)
+
+				kafka, err := client.Services.Get(project, serviceName)
+				if err != nil {
+					return false
+				}
+
+				if v, ok := kafka.UserConfig["kafka_version"]; ok {
+					if version, err := strconv.ParseFloat(v.(string), 64); err == nil {
+						if version >= 3 {
+							return true
+						}
+					}
+				}
+
+				return false
+			}),
+		),
 	}
 }
 
@@ -119,4 +160,48 @@ func resourceKafkaCreate(ctx context.Context, d *schema.ResourceData, m interfac
 	}
 
 	return nil
+}
+
+func resourceKafkaRead(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
+	client := m.(*aiven.Client)
+
+	kafka, err := client.Services.Get(splitResourceID2(d.Id()))
+	if err := resourceReadHandleNotFound(err, d); err != nil {
+		return nil
+	}
+
+	var diags diag.Diagnostics
+	var kafkaVersion float64
+	var schemaRegistry bool
+	var kafkaRest bool
+
+	if v, ok := kafka.UserConfig["kafka_version"]; ok {
+		if version, err := strconv.ParseFloat(v.(string), 64); err == nil {
+			kafkaVersion = version
+		}
+	}
+
+	if v, ok := kafka.UserConfig["schema_registry"]; ok && v.(bool) {
+		schemaRegistry = true
+	}
+
+	if v, ok := kafka.UserConfig["kafka_rest"]; ok && v.(bool) {
+		kafkaRest = true
+	}
+
+	// Checking is Confluent SR/REST -> Karapace migration is available
+	if kafkaVersion < 3.0 &&
+		((schemaRegistry && !kafka.Features.Karapace) ||
+			(kafkaRest && !kafka.Features.KafkaRest)) {
+		diags = append(diags, diag.Diagnostic{
+			Severity: diag.Warning,
+			Summary: "You are using Confluent Schema Registry v5.0 that is no longer supported " +
+				"on Kafka v3.0. Please switch to Karapace, a drop-in open source replacement " +
+				"before proceeding with the upgrade. To do that use aiven_kafka.karapace=true " +
+				"that will switch the service to use Karapace for schema registry and REST proxy. " +
+				"For more information, please refer to our help article: https://help.aiven.io/en/articles/5651983",
+		})
+	}
+
+	return append(diags, resourceServiceRead(ctx, d, m)...)
 }